Op.25 No.8 proofreading
Measure 3, beat 3, right hand: In the second chord in the triplet,
replace the G-flat/E-flat with E-flat/C-natural to continue the
chromatic sequence.

Measure 5, beat 2, left hand: Replace the F-natural with A-flat in the
lower note of the first chord.

Measure 10, beat 4, left hand: Replace the D-flat with D-natural in the
lower note of the last chord.

Measure 14, beats 1 and 2, right hand: The sequence of 6 chords should
begin with E-flat/C-natural instead of D-flat/B-flat (like the sequence
from measure 13). This aligns with the right hand figuration pattern in
the following two measures, 15 and 16.

Measure 17, beat 3, right hand: To align with the sequence in progress,
replace the last chord in the triplet (D-flat/B-flat) with
E-flat/C-natural.